refactor(api): name fetch-season-games after its export; consolidate bracket work into the MM epic
Feedback items 3 and 5.

Every sibling in `services/nfl/` matches its filename to its main export;
`fetch-week-games.ts` exporting `fetchSeasonGames` was the outlier. The
function name is the accurate one — it fetches a season's weeks and dedupes
to one game per id — so the file moves.

Bracket work now lives entirely in the March Madness epic: SIM-8 becomes
MM-8, with its dependency corrected from MM-1 (settings schema, which
already shipped inside LG-2 and is now marked done) to MM-2, the bracket
ingestion that actually introduces the NCAAMB provider surface. MVP scope is
unchanged — March Madness stays an MVP mode, sequenced last.
