+++ apps/api/src/services/nfl/fetch-season-games.ts
